Subiendo el Backend para Render

+++ src/main/java/com/portfolio/argprograma/repository/RBanner.java
@@ -9,6 +9,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RBanner extends JpaRepository<Banner, Long> {
+public interface RBanner extends JpaRepository<Banner, Integer> {
     
 }
