+++ src/main/java/com/portfolio/argprograma/repository/REducacion.java
@@ -9,6 +9,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface REducacion extends JpaRepository<Educacion, Long>{
+public interface REducacion extends JpaRepository<Educacion, Integer>{
     
 }
